+++ shared/src/androidMain/kotlin/com/tsymbaliuk/currency/converter/network/datasource/FirebaseNetworkDataSource.kt
@@ -1,9 +1,9 @@
-package com.tsymbaliuk.currency.converter
+package com.tsymbaliuk.currency.converter.network.datasource
 
 import com.google.firebase.ktx.Firebase
 import com.google.firebase.remoteconfig.ktx.remoteConfig
 
-actual class RemoteConfig {
+actual class FirebaseNetworkDataSource {
 
     init {
         Firebase.remoteConfig.fetchAndActivate()
