--- shared/src/commonMain/kotlin/com/tsymbaliuk/currency/converter/RemoteConfig.kt
@@ -1,8 +1,0 @@
-package com.tsymbaliuk.currency.converter
-
-expect class RemoteConfig() {
-
-    fun getApiKey(parameterName: String = NetworkConstants.FIREBASE_REMOTE_CONFIG_API_KEY): String
-
-}
-
