--- shared/src/iosMain/kotlin/com/tsymbaliuk/currency/converter/RemoteConfig.kt
@@ -1,9 +1,0 @@
-package com.tsymbaliuk.currency.converter
-
-actual class RemoteConfig {
-
-    actual fun getApiKey(parameterName: String): String {
-        return "test"
-    }
-
-}
