+++ shared/src/iosMain/kotlin/com/tsymbaliuk/currency/converter/network/datasource/FirebaseNetworkDataSource.kt
@@ -1,0 +1,9 @@
+package com.tsymbaliuk.currency.converter.network.datasource
+
+actual class FirebaseNetworkDataSource {
+
+    actual fun getApiKey(parameterName: String): String {
+        return "test"
+    }
+
+}
